投稿者のみ削除が行えるよう条件分岐

+++ app/controllers/items_controller.rb
@@ -26,8 +26,12 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+      if item.user == current_user
+        item.destroy
+        redirect_to root_path
+      else
+        root_path
+      end
   end
 
   private
